Data Refactor Start + Quiz Page Updates
Need to look over my logic blocks and make sure the loading and saving of data is easy and clean. Right now its a bit of a nightmare and needs a bit of a refactor. (did a bit of this on Visits and Quizes, need to extend to all data loading and saving)